requirements.txt
h11~=0.14.0
pillow~=11.1.0
typing_extensions~=4.13.1
sniffio~=1.3.1
certifi~=2025.1.31
fastapi~=0.115.12
starlette~=0.46.1
pydantic~=2.11.3
pydantic_core~=2.33.1
requests~=2.32.3
anyio~=4.9.0
urllib3~=2.3.0
click~=8.1.8
uvicorn~=0.34.0
annotated-types~=0.7.0
idna~=3.10
charset-normalizer~=3.4.1
python-multipart~=0.0.20
pywinstyles~=1.8